fix: Only enable enrollment report emails if the configuration allows
If a Tutor config.yml file sets "RUN_SMTP: false" and does not set
SMTP_HOST, then SMTP_HOST is not available to the Jinja2 templates,
and "tutor config save" fails.

Test for those two variables, and if neither are set, set the
enrollment_report_mail_enable parameter to false.