fix: Remove SMTP_HOST from ENROLLMENTREPORTS_MAIL_TO default
If SMTP_HOST is unset in the configuration, then its default ("smtp")
is available to Tutor's Jinja2 templates, but not to a plugin add a
variable. Thus, having it in the default when SMTP_HOST is unset will
cause "tutor config save" to fail. We don't want that, so set the
ENROLLMENTREPORTS_MAIL_TO default to the empty list instead.