fix: Correct test for SMTP_USERNAME
Tutor sets SMTP_USERNAME to the empty string by default, so it is
always defined. Correct the check by testing for "if SMTP_USERNAME"
(which evaluates to false if it is set to the empty string) rather
than "is SMTP_USERNAME is defined" (which evaluates to true in that
case).